Attempt to load secret key from environment before failing (#52)
Normalizes the behaviour of secret keys, either loaded from cloud
profiles or via environment variable. If we receive a secret key error
in the explicit options phase (parse error, missing key), we should not
expose that error until we've had a chance to parse the environment
variables and potentially get ourselves a valid secret key.

Previously we special-cased the cloud profile so that it was consulted
first. This caused https://github.com/geldata/gel-cli/issues/1590
because of the unclear priority.

This should not affect any working use cases -- it should make things
somewhat easier to work with.